extract methods for reuse
